Move orders sync-schema to separate file
