+++ workers/loc.api/sync/schema/sync-schema/index.js
@@ -13,6 +13,7 @@ const trades = require('./trades')
 const fundingTrades = require('./funding-trades')
 const publicTrades = require('./public-trades')
 const statusMessages = require('./status-messages')
+const orders = require('./orders')
 
 const _methodCollMap = new Map([
   [SYNC_API_METHODS.LEDGERS, ledgers],
@@ -20,21 +21,7 @@ const _methodCollMap = new Map([
   [SYNC_API_METHODS.FUNDING_TRADES, fundingTrades],
   [SYNC_API_METHODS.PUBLIC_TRADES, publicTrades],
   [SYNC_API_METHODS.STATUS_MESSAGES, statusMessages],
-  [
-    SYNC_API_METHODS.ORDERS,
-    {
-      name: ALLOWED_COLLS.ORDERS,
-      maxLimit: 2500,
-      dateFieldName: 'mtsUpdate',
-      symbolFieldName: 'symbol',
-      sort: [['mtsUpdate', -1]],
-      hasNewData: false,
-      start: [],
-      isSyncRequiredAtLeastOnce: false,
-      type: COLLS_TYPES.INSERTABLE_ARRAY_OBJECTS,
-      model: getModelOf(TABLES_NAMES.ORDERS)
-    }
-  ],
+  [SYNC_API_METHODS.ORDERS, orders],
   [
     SYNC_API_METHODS.MOVEMENTS,
     {
